moved providers to project root
