+++ app/layout.tsx
@@ -4,8 +4,8 @@ import type { PropsWithChildren } from "react"
 import { SpeedInsights } from "@vercel/speed-insights/next"
 import Footer from "@/components/footer"
 import Navbar from "@/components/navbar"
-import { ThemeProvider } from "@/components/providers/theme-provider"
-import { CvDataStoreProvider } from "@/components/providers/cv-data-store-provider"
+import { ThemeProvider } from "@/providers/theme-provider"
+import { CvDataStoreProvider } from "@/providers/cv-data-store-provider"
 import { Toaster } from "@/components/ui/toaster"
 import { siteConfig } from "@/config/site"
 import { fontDisplay, fontSans } from "@/lib/fonts"
