+++ app/show/page.tsx
@@ -4,7 +4,7 @@ import type { CvDataStore } from "@/lib/stores/cv-data-store.types"
 import { Button } from "@/components/ui/button"
 import Loader from "@/components/ui/loader"
 import { useAsyncErrors } from "@/hooks/use-async-errors"
-import { useCvDataStore } from "@/components/providers/cv-data-store-provider"
+import { useCvDataStore } from "@/providers/cv-data-store-provider"
 import { format } from "date-fns"
 import dynamic from "next/dynamic"
 import { useEffect, useState } from "react"
