+++ components/email-form.tsx
@@ -3,7 +3,7 @@ import { Button } from "@/components/ui/button"
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
 import { emailSchema } from "@/form-generator/validation-schemas"
-import { useCvDataStore } from "@/components/providers/cv-data-store-provider"
+import { useCvDataStore } from "@/providers/cv-data-store-provider"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useRouter } from "next/navigation"
 import { useForm } from "react-hook-form"
