+++ providers/theme-provider.tsx
@@ -1,8 +1,8 @@
 "use client"
+import type { ThemeProviderProps } from "next-themes/dist/types"
 
 import * as React from "react"
 import { ThemeProvider as NextThemesProvider } from "next-themes"
-import { type ThemeProviderProps } from "next-themes/dist/types"
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
   return <NextThemesProvider {...props}>{children}</NextThemesProvider>
